requirements.txt
streamlit
langchain
langchain-aws
langchain-community
python-dotenv
boto3
botocore
pypdf
faiss-cpu
PyCryptodome
tiktoken
gpt4all
